test(maps): move map tests into nested directories
