+++ tests/unit/maps/SymbolMap/SymbolMap.spec.ts
@@ -1,14 +1,14 @@
 import { promises as fs } from 'fs'
 import { join, resolve } from 'path'
 import { shallowMount } from '@vue/test-utils'
-import SymbolMap from '@/maps/SymbolMap.vue'
+import SymbolMap from '@/maps/SymbolMap/SymbolMap.vue'
 
 vi.mock('d3', async () => {
   return {
     ...(await vi.importActual('d3')),
     json: async (url) => {
       const pathname = url.split('https://icij.github.io/murmur/').pop()
-      const abspath = resolve(__dirname, join('../../../public', pathname))
+      const abspath = resolve(__dirname, join('../../../../public', pathname))
       const raw = await fs.readFile(abspath, 'UTF-8')
       return JSON.parse(raw)
     }
